Tweak the navbar-brand a bit more.
Instead of adding opacity to the navbar brand as a whole, this only adds
opacity to the background color.  This doesn't change anything in the
math4, math4-green, and math4-red themes as the text is white anyway,
but makes a big difference in the math4-yellow theme in which the text
is black.  The opacity on the text is not good there.

This also adds a 1 pixel border on the bottom of the navbar-brand.  This
gives a nicer division between the navbar-brand and the site nav
elements below.  Particularly when the top item is active.